Consolidate assets to SDK, import from SDK
