+++ packages/demo/backend/src/config/markets.ts
@@ -1,8 +1,13 @@
-import { ETH, type LendMarketConfig, WETH } from '@eth-optimism/actions-sdk'
+import {
+  ETH,
+  type LendMarketConfig,
+  USDC,
+  WETH,
+} from '@eth-optimism/actions-sdk'
 import type { Address } from 'viem'
 import { baseSepolia, optimismSepolia, unichain } from 'viem/chains'
 
-import { USDC, USDC_DEMO } from './assets.js'
+import { USDC_DEMO } from './assets.js'
 
 export const GauntletUSDC: LendMarketConfig = {
   address: '0x38f4f3B6533de0023b9DCd04b02F93d36ad1F9f9' as const,
