+++ packages/demo/frontend/src/constants/markets.ts
@@ -1,23 +1,13 @@
 import {
   ETH,
   type LendMarketConfig,
+  USDC_DEMO,
   WETH,
 } from '@eth-optimism/actions-sdk/react'
-import type { Asset } from '@eth-optimism/actions-sdk/react'
 import type { Address } from 'viem'
 import { baseSepolia, optimismSepolia } from 'viem/chains'
 
-export const USDC_DEMO: Asset = {
-  address: {
-    [baseSepolia.id]: '0xb1b0FE886cE376F28987Ad24b1759a8f0A7dd839',
-  },
-  metadata: {
-    decimals: 6,
-    name: 'USDC',
-    symbol: 'USDC_DEMO',
-  },
-  type: 'erc20',
-}
+export { USDC_DEMO }
 
 export const GauntletUSDCDemo: LendMarketConfig = {
   address: '0x018e22BBC6eB3daCfd151d1Cc4Dc72f6337B3eA1' as const,
