+++ packages/sdk/src/index.ts
@@ -1,5 +1,5 @@
 export type { Actions } from '@/actions.js'
-export { ETH, USDC, WETH } from '@/constants/assets.js'
+export { ETH, MORPHO, USDC, USDC_DEMO, WETH } from '@/constants/assets.js'
 export type { SupportedChainId } from '@/constants/supportedChains.js'
 export * from '@/core/error/errors.js'
 export { LendProvider, MorphoLendProvider } from '@/lend/index.js'
